Encapsulate copy code button to new component
