+++ packages/docusaurus-theme-classic/src/theme/CodeBlock/index.tsx
@@ -8,8 +8,6 @@
 import React, {isValidElement, useEffect, useState} from 'react';
 import clsx from 'clsx';
 import Highlight, {defaultProps, type Language} from 'prism-react-renderer';
-import copy from 'copy-text-to-clipboard';
-import {translate} from '@docusaurus/Translate';
 import {
   useThemeConfig,
   parseCodeBlockTitle,
@@ -18,6 +16,7 @@ import {
   ThemeClassNames,
   usePrismTheme,
 } from '@docusaurus/theme-common';
+import CopyButton from '@theme/CodeBlock/CopyButton';
 import type {Props} from '@theme/CodeBlock';
 
 import styles from './styles.module.css';
@@ -31,7 +30,6 @@ export default function CodeBlock({
 }: Props): JSX.Element {
   const {prism} = useThemeConfig();
 
-  const [isCopied, setIsCopied] = useState(false);
   const [mounted, setMounted] = useState(false);
   // The Prism theme on SSR is always the default theme but the site theme
   // can be in a different mode. React hydration doesn't update DOM styles
@@ -90,15 +88,6 @@ export default function CodeBlock({
     languageProp ?? parseLanguage(blockClassName) ?? prism.defaultLanguage;
   const {highlightLines, code} = parseLines(content, metastring, language);
 
-  const handleCopyCode = () => {
-    copy(code);
-    setIsCopied(true);
-
-    setTimeout(() => {
-      setIsCopied(false);
-    }, 2000);
-  };
-
   return (
     <Highlight
       {...defaultProps}
@@ -153,38 +142,7 @@ export default function CodeBlock({
               </code>
             </pre>
 
-            <button
-              type="button"
-              aria-label={
-                isCopied
-                  ? translate({
-                      id: 'theme.CodeBlock.copied',
-                      message: 'Copied',
-                      description: 'The copied button label on code blocks',
-                    })
-                  : translate({
-                      id: 'theme.CodeBlock.copyButtonAriaLabel',
-                      message: 'Copy code to clipboard',
-                      description: 'The ARIA label for copy code blocks button',
-                    })
-              }
-              className={clsx(
-                styles.copyButton,
-                'clean-btn',
-                isCopied && styles.copyButtonCopied,
-              )}
-              onClick={handleCopyCode}>
-              <span className={styles.copyButtonIcons} aria-hidden="true">
-                <svg className={styles.copyButtonIcon} viewBox="0 0 24 24">
-                  <path d="M19,21H8V7H19M19,5H8A2,2 0 0,0 6,7V21A2,2 0 0,0 8,23H19A2,2 0 0,0 21,21V7A2,2 0 0,0 19,5M16,1H4A2,2 0 0,0 2,3V17H4V3H16V1Z" />
-                </svg>
-                <svg
-                  className={styles.copyButtonSuccessIcon}
-                  viewBox="0 0 24 24">
-                  <path d="M21,7L9,19L3.5,13.5L4.91,12.09L9,16.17L19.59,5.59L21,7Z" />
-                </svg>
-              </span>
-            </button>
+            <CopyButton code={code} />
           </div>
         </div>
       )}
