+++ packages/docusaurus-theme-classic/src/theme/CodeBlock/styles.module.css
@@ -43,66 +43,6 @@
   border-radius: var(--ifm-global-radius);
 }
 
-.copyButton {
-  display: flex;
-  background: inherit;
-  border: 1px solid var(--ifm-color-emphasis-300);
-  border-radius: var(--ifm-global-radius);
-  padding: 0.4rem;
-  position: absolute;
-  right: calc(var(--ifm-pre-padding) / 2);
-  top: calc(var(--ifm-pre-padding) / 2);
-  transition: opacity 200ms ease-in-out;
-  opacity: 0;
-}
-
-.copyButton:focus-visible,
-.copyButton:hover,
-.codeBlockContainer:hover .copyButtonCopied {
-  opacity: 1 !important;
-}
-
-.codeBlockContainer:hover .copyButton:not(.copyButtonCopied) {
-  opacity: 0.4;
-}
-
-.copyButtonIcons {
-  position: relative;
-  width: 1.125rem;
-  height: 1.125rem;
-}
-
-.copyButtonIcon,
-.copyButtonSuccessIcon {
-  position: absolute;
-  top: 0;
-  left: 0;
-  fill: currentColor;
-  opacity: inherit;
-  width: inherit;
-  height: inherit;
-  transition: all 0.15s ease;
-}
-
-.copyButtonSuccessIcon {
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%) scale(0.33);
-  opacity: 0;
-  color: #00d600;
-}
-
-.copyButtonCopied .copyButtonIcon {
-  transform: scale(0.33);
-  opacity: 0;
-}
-
-.copyButtonCopied .copyButtonSuccessIcon {
-  transform: translate(-50%, -50%) scale(1);
-  opacity: 1;
-  transition-delay: 0.075s;
-}
-
 .codeBlockLines {
   font: inherit;
   /* rtl:ignore */
